+++ releases.md
@@ -1,5 +1,21 @@
 ## Releases
 
+### v0.1.2
+Release Date: 2025-feb-28
+
+- Still incomplete and not fully working
+- Directory objects are mostly working but still many bugs wit cache
+- Resource objects are still to be migrated
+- cmd/azm:
+  - Continuing to migrate old azm options to this new version 
+- pkg/maz recent changes:
+  - Simplified GetMatchingObjects()
+  - Upgraded package dependencies:
+    - github.com/queone/utl v1.3.1
+  - Dropped ApiErrorCheck() and embeded the error checking directly into ApiCall()
+  
+---
+
 ### v0.1.1
 Release Date: 2025-feb-23